Courtyard
Brandon's Courtyard map.
I added simple flowerbeds for the flowers to go in, and a few vases to populate the room. Also changed the walls slightly so they fit alang the front wall, and put some (placeholder) particles on the fountain.
Added some post-processing volumes to make the past and future look more different.